requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
fpdf==1.7.2
idna==3.11
numpy==2.4.2
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
urllib3==2.6.3
